Factor url related functions in urlState service
